environment and agents reworked
